+++ src/graphql/queries.ts
@@ -107,6 +107,7 @@ export type NetworkCostsData = {
 export type NetworkCostsFilter = {
   networkLevel: Scalars["String"]["input"];
   operatorId: Scalars["Int"]["input"];
+  operatorOnly?: InputMaybe<Scalars["Boolean"]["input"]>;
   period: Scalars["Int"]["input"];
 };
 
@@ -473,6 +474,7 @@ export type TariffsData = {
 export type TariffsFilter = {
   category: Scalars["String"]["input"];
   operatorId: Scalars["Int"]["input"];
+  operatorOnly?: InputMaybe<Scalars["Boolean"]["input"]>;
   period: Scalars["Int"]["input"];
 };